requirements.txt
PyJWT
pymongo
celery
Flask